==> notebooks/1. extract_entities.ipynb <==
"""
1. Extract a list of entites from MP's written questions
"""

"""
SPARQL query to api.parliament.uk endpoint to get all written questions

Altered so it does not contain questions from the House of Lords
"""


# Help setting this up from https://rebeccabilbro.github.io/sparql-from-python/
from SPARQLWrapper import SPARQLWrapper, JSON
import pandas as pd

question_df = pd.DataFrame()
sparql  = SPARQLWrapper("https://api.parliament.uk/sparql/")
sparql.setQuery("""
SELECT *
WHERE {
?question <https://id.parliament.uk/schema/writtenQuestionIndexingAndSearchUin> ?qnum .
?person <https://id.parliament.uk/schema/askingPersonHasQuestion> ?question .
?person <https://id.parliament.uk/schema/wikidataThingHasEquivalentWikidataResource> ?wikidataperson.
?question <https://id.parliament.uk/schema/questionText> ?text .
?question <https://id.parliament.uk/schema/questionAskedAt> ?date .
FILTER (?date >= "2023-01-01+00:00"^^xsd:dateTime && ?date < "2023-10-01+00:00"^^xsd:dateTime)
FILTER regex(?qnum, "^(?!HL)") 
}
""")

sparql.setReturnFormat(JSON)
result = sparql.query().convert()
# results -> bindings returns a list of dictionaries
result = result["results"]["bindings"]

# removed type from the results
for item in result:
    for key in item:
        item[key] = item[key]["value"]


question_df = pd.DataFrame(result)

# remove tags from text
question_df["text"] = question_df["text"].str.replace("<p>", "", )
question_df["text"] = question_df["text"].str.replace("</p>", "", )

question_df.head(3)
question_df.to_csv("../data/questions.csv", index=False)
len(question_df)

"""
Functions for named entity recognition and disambiguation
"""

import requests
import random
token = "5fcf52fb-79b2-4718-b7ee-62957df7d7e0-843339462"

def pick_random_question():
    x = random.randint(0, len(question_df))
    return question_df["text"].iloc[x]


def tag_me(text=None):
    if not text:
        text = pick_random_question()
        print(text)


    address = f"https://tagme.d4science.org/tagme/tag?lang=en&gcube-token={token}&include_categories=false&include_abstract=false&text={text}"
    response = requests.get(address)

    annotations = response.json()["annotations"]
    new_annotations = []

    for spot in annotations:
        # remove if rho is less than 0.1
        if spot["rho"] >= 0.1:
            new_annotations.append(spot)

    return new_annotations


def WAT(text=None):
    if not text:
        text = pick_random_question()
        #print(text)
       
    address = f"https://wat.d4science.org/wat/tag/tag?lang=en&gcube-token={token}&include_categories=true&text={text}"
    response = requests.get(address)
    annotations = response.json()["annotations"]
    new_annotations = []

    for spot in annotations:
        # remove if rho is less than 0.1
        if spot["rho"] >= 0.1:
            new_annotations.append(spot)
    return new_annotations


def REL(text=None):
    if not text:
        text = pick_random_question()
        #print(text)
    API_URL = "https://rel.cs.ru.nl/api"

    # Example ED.
    ed_result = requests.post(API_URL, json={
        "text": text,
    })
    assert ed_result.status_code == 200
    
    return ed_result.json()

# Time how long it takes for each call to WAT vs REL
import time
iters = 50


start = time.time()
for _ in range(iters):
    WAT()
end = time.time()
print("WAT average time: ", (end-start)/iters)


start = time.time()
for _ in range(iters):
    REL()
end = time.time()
print("REL average time: ", (end-start)/iters)    

"""
Functions for getting a location from a wikipedia page and converting it to British National Grid units (Easting, Northing)
from longitude and latitude
"""

# Help getting wikipedia coordinates from https://stackoverflow.com/questions/40098656/how-to-get-coordinates-from-a-wikipedia-page-through-api
def get_wikipedia_location(wikipedia_title):
    response = requests.get(f"https://en.wikipedia.org/w/api.php?action=query&prop=coordinates&format=json&titles={wikipedia_title}").json()
    try :
        # makes iterator over pages and takes first element
        page = next(iter(response["query"]["pages"].values()))
        return page["coordinates"][0]["lat"], page["coordinates"][0]["lon"]
    except KeyError:
        return None

def convert_to_BNG(location):
    result = requests.get(f"http://webapps.bgs.ac.uk/data/webservices/CoordConvert_LL_BNG.cfc?method=LatLongToBNG&lat={location[0]}&lon={location[1]}").json()
    return result["EASTING"], result["NORTHING"]


"""
Iterates through the stored csv file and combines them into one dataframe
"""
import os
def import_question_location_chunks():

    question_locations_df = pd.DataFrame()
    for file in os.listdir("../question_entities"):
        if file.endswith(".csv"):
            print(file)
            new_questions = pd.read_csv(f"../question_entities/{file}")
            # reset the index

            new_questions = new_questions.drop(columns=["Unnamed: 0"])
            question_locations_df = pd.concat([question_locations_df, new_questions], ignore_index=True)
    return question_locations_df

"""
Iterates through each question, extracts entities and then finds their location in BNG.

question_locations stores the question, entity, easting, northing and wikipedia link.

This is then stored in a dataframe and saved to a csv file, so it can be performed in chunks.
"""
import time
import os
start = 0
end = 0


if os.path.exists("../question_entities/question_locations_*.csv"):
    question_locations_df = import_question_location_chunks()
else:
    banned_entities = [ "United_Kingdom", "England", "Wales", "Northern_Ireland", "Scotland"]
    question_locations = []

    for index, row in question_df[start:end].iterrows():
        for entity in WAT(row["text"]):
            while True:
                try:
                    if entity["title"] in banned_entities:
                        break

                    location = get_wikipedia_location(entity["title"])
                    # sleep for 0.1 seconds to avoid rate limiting
                    
                    if location is not None:
                        location = convert_to_BNG(location)

                        # origin is in the isles of scilly, and spans from 0 to 700,000 easting and 0 to 1,300,000 northing
                        if location[0] >= 0 and location[0] <= 700_000 and location[1] >= 0 and location[1] <= 1_300_000:
                            print(entity["title"])
                            question_locations.append({"question": row["question"], "entity": entity["title"], "easting": location[0], "northing": location[1], "wikipedia": f"https://en.wikipedia.org/wiki/{entity["title"]}"})
                    break
                except KeyError:
                    print("time out")
                    time.sleep(3)

    question_locations_df = pd.DataFrame(question_locations)
    question_locations_df.to_csv(f"../question_entities/question_locations_{end}.csv")
    question_locations_df

# prints out the most commonly occuring entity and a question it appears in
for entity in question_locations_df["entity"].value_counts().index:
    print(entity + " : " + str(question_locations_df[question_locations_df["entity"] == entity]["question"].iloc[0]))

# these entities are removed as they are there location is not relevant when a given question is asked
remove_entities = ["Office_for_National_Statistics", "Department_for_Transport", "HM_Treasury", "Cabinet Office", "House_of_Commons_of_the_United_Kingdom", "Parliament_of_the_United_Kingdom", "National_Audit_Office_(United_Kingdom)", "Government_Equalities_Office", "Cabinet_Office", "London", "Ofcom"]
question_locations_df = question_locations_df[~question_locations_df["entity"].isin(remove_entities)]


# save the dataframe to a csv file
question_locations_df.to_csv("../data/question_locations_full.csv", index=False)
question_locations_df
