==> notebooks/2. link_MPs_to_geometry.ipynb <==
"""
2. Link MP's to constituencies and their geometry
"""

"""
Produces dataframe of geometry of all westminister constituencies
"""
import pandas as pd
import geopandas as gpd
en_wl_sct = gpd.read_file("../shape/westminster_const_region.shp")
ni = gpd.read_file("../shape/nireland_aa_2008.shp")

en_wl_sct = en_wl_sct.rename(columns={"NAME": "name"})
uk_geometry = pd.concat([en_wl_sct[["name", "geometry"]], ni[["name", "geometry"]]])

"""
SPARQL query to wikidata endpoint to get all MPs and their constituencies

"""
from SPARQLWrapper import SPARQLWrapper, JSON

# Help setting this up from https://rebeccabilbro.github.io/sparql-from-python/
const_df = pd.DataFrame()
sparql  = SPARQLWrapper("https://query.wikidata.org/sparql")
sparql.setQuery("""
SELECT ?person ?const ?personLabel ?constLabel
WHERE
{
  ?person p:P39 ?position.
  ?position ps:P39 wd:Q77685926.
  ?position pq:P768 ?const.
  SERVICE wikibase:label { bd:serviceParam wikibase:language "en". }
}
""")

sparql.setReturnFormat(JSON)
result = sparql.query().convert()
# results -> bindings returns a list of dictionaries
result = result["results"]["bindings"]

# removed type from the results
for item in result:
    for key in item:
        item[key] = item[key]["value"]
const_df = pd.DataFrame(result)
const_df.head(3)

# remove duplicate records (which can occur when a person changes parties i.e Julian Lewis)
const_df = const_df.drop_duplicates()
const_df = const_df.astype({"constLabel": "object", "personLabel": "object"})

# I have decided not remove duplicate people for the same constituency, which can happen if a bi-election occurs.
# This is because either of them could have asked questions.

"""
Makes the names of the constituencies consistent between uk_geometry and const_df
"""
replace_dict = {
    "," : "",
    "Co Const": "",
    "Cor Const" : "",
    "Boro Const" : "",
    "Burgh Const" : "",
    "." : "",
    "Weston-Super-Mare" : "Weston-super-Mare",
    "ô" : "o"
}

for key in replace_dict:
    uk_geometry["name"] = uk_geometry["name"].str.replace(key, replace_dict[key])
    const_df["constLabel"] = const_df["constLabel"].str.replace(key, replace_dict[key])

# remove whitespace that is at the end of some names
uk_geometry["name"] = uk_geometry["name"].str.rstrip()

uk_geometry = uk_geometry[["name", "geometry"]]

MP_geometry = pd.merge(const_df, uk_geometry, left_on="constLabel", right_on="name", how="inner")
print(len(MP_geometry["constLabel"].unique()))

"""
Check for any constituencies present in one but not the other
"""
for name in uk_geometry["name"].unique():
    if name not in MP_geometry["constLabel"].unique():
        print(name)

for district in const_df["constLabel"].unique():
    if name not in MP_geometry["name"].unique():
        print(name)

MP_geometry.drop(columns=["name"], inplace=True)
MP_geometry

# save to csv
MP_geometry.to_csv("../data/MP_geometry.csv", index=False)

MP_geometry.info()
