==> notebooks/3. link_entities_to_const.ipynb <==
'''
3. Relate entities in MP's questions to constituencies
'''

import pandas as pd
import geopandas as gpd

MP_geometry = pd.read_csv("../data/MP_geometry.csv")

question_locations = pd.read_csv("../data/question_locations_full.csv")

# converts MP_geometry geometry column to type geometry
MP_geometry['geometry'] = gpd.GeoSeries.from_wkt(MP_geometry['geometry'])
MP_geometry.info()

'''
Checks for each entity which constituency it belongs and then adds it to the new constituency column
'''
import shapely
entity_const = question_locations.copy()
# help from https://stackoverflow.com/questions/7861196/check-if-a-geopoint-with-latitude-and-longitude-is-within-a-shapefile

# adds a new column to question_locations_df
entity_const["const"] = None


for index, row in entity_const.iterrows():
    point = shapely.geometry.Point(row["easting"], row["northing"])
    for index2, row2 in MP_geometry.iterrows():
        if row2["geometry"].contains(point):
            print(index)
            entity_const.at[index, "const"] = row2["constLabel"]
            break

# show entities that do not have a constituency
# these are mainly rivers or where wikipedia has put the coordinates in the sea
entity_const[entity_const["const"].isnull()]

entity_const[entity_const["const"].isnull()]["entity"].unique()

import numpy as np

# pick closest entites for these
for index, row in entity_const[entity_const["const"].isnull()].iterrows():
    shortest_dist = np.inf
    
    # help from https://stackoverflow.com/questions/36972537/distance-from-point-to-polygon-when-inside

    point = shapely.geometry.Point(row["easting"], row["northing"])
    for index2, row2 in MP_geometry.iterrows():
        dist = row2["geometry"].boundary.distance(point)

        if dist < shortest_dist:
            shortest_dist = dist
            closest_const = row2["constLabel"]
    entity_const.at[index, "const"] = closest_const

entity_const["entity"].value_counts() 

for index, row in entity_const.iterrows():
    print(row["entity"])

entity_const.to_csv("../data/entity_const.csv", index=False)
